Created job scheduler and Job class
The job scheduler runs in a single seperate thread as a daemon. The
daemon attribute should ensure the thread is killed when the main
program terminates.

The job scheduler takes a subclass of the Job class. These subclasses require
a function called run() to be present which will be called from the
scheduler.