Added initial controller and client network code
Created the initial controller based on my code of the publish/subscribe
project. It uses python's asyncio library for handling asynchronous
connections from nodes.

Also created the initial network client code that will be used to
connect to the controller. We will be using JSON messages between the
controller and nodes to send messages. Not using asyncio here but
nonblocking sockets for the moment. (I think the message buffer is not
working yet but rather we're taking data form the socket and only
process the first received message. Code was written a while back
though so might be mistaken!!)

A node will have to first register itself to the controller. The
controller will confirm the registration to the node after which the
node can send probe result data to the controller for processing. A
heartbeat will also have to be implemented between the node and
controller to ensure everything is still functioning properly.

Next task will be to somehow get the results from scheduled jobs passed
to the node class which can in turn send the result to the controller