Quick hack to get probe results to controller
I'm passing the controller session to the job_scheduler. From the
scheduler we're calling the probe_result method to send the result to
the controller. At the moment I'm just pasting the probe result as a
list. The probe results aren't standarised so looks ugly but at least we
have something up and running now!