Renamed node connection manager classes
Now using the class names MessageHandler and ConnectionProtocol
instead of Connector and ConnectorProtocol. This is hopefully a more
descriptive name of the task they perform.

In addition we're now excluding all log.* messages from git.

Also created dummy files for some of the required modules like
the configuration manager and service monitor