Task Manager to use tcp conn to asyncio controller
Still trying to come up with a good architecture of the controller and
node components. I've removed the asyncio based task_manager and
replaced it with the code from the JobScheduler written earlier for the
node. The idea is now that the task_manager will use a tcp_client
connection towards the asyncio node controller server to send and recv
messages between the various nodes.

The tasks.py file will be containing any specific tasks that are
subclassed from task_manager.Task. They will at minimum have to
implement def run() as this will be used by the task_manager to launch
the task. At the moment there is only a single Task defined there which
is RegisterNode. This will be used to register the node to the asyncio
node controller server although at the moment it's only printing out a
text string.

lastly renamed the node_connector to tcp_server