Task Manager registering to controller tcp server
Since we're running the task manager loop in a seperate thread I couldn't
pass the message_handler function from main.py. If I did this any
changes on the message_handler variables we're not passed onto the
main.py thread.

I'm now setting up the message_handler in the task_manager only. This
should be enough as the main.py thread would not need to directly
interface with it. We will use the main.py event loop to submit probes to
the task_manager

Next step is to gracefully disconnect a session when the main program
ends