Added unregister function to message_handler
The message_handler now supports the unregister function. This will
remove a node from the registered node list. UnregisterTask class is
created that can be used by the task_manager. When the
task_manager.stop() method is called we schedule this UnregisterTask to
gracefully disconnect from the controller.

Eventually we want to implement a healthcheck between nodes and
controllers to catch any disconnected sessions.