Added run_at and recurrence to Task
The __init__ function of the Task class now allows you to tell when to
run the task and set task recurrence if any. The sub classes should not
overwrite the __init__ class or at least call the base __init__ class
function if they have to. Also we're now setting the Task self.name
variable to the class name. We can use this name to find out the Task
type for usage in task_manager