Added first (dummy) probe
Added our first dummy probe called IcmpProbe. It requires the dest_addr
argument to be passed and uses the super().__init__() to inherit all the
init functionality of the main class Task. Still have to verify this
actually works the way i expect. I think i'm still able to set the
run_as etc variables on the new class because of the super() call even
though I'm not mentioning those variables in the new class.