Implemented task recurrence functionality
Now able to schedule a recurring task providing the recurrence_time in
seconds and recurrence_count. Created a new function in the base Task
class that checks if the task has to be rescheduled again. The
task_manager calls this function and if true adds the task back to the
queue.