Added send capability to tcp_server
The tcp_server is now able to send messages back to the client. To test
functionality we're now sending a reply back when we receive a message
with 'type': 'probe'.

At the moment the task_manager is calling the
message_handler.recv_message and then checking down to the tcp_client
protocol level if there's a message in the queue. Will probably have to
refactor this code to get this handled in the message_handler section as
we don't really want to dig into the protocol level from the
task_manager.