Created TraceProbe Task
The TraceProbe task is using the underlying os traceroute command and
parsing it's output. At the moment it prints out a list containing the
hops and rtt. It's based on a script i wrote at work that does the same
but resolves the hops against our IP address management tool to get
additonal details on the hops (network, hostname, ticket number,
comments, etc)