Added ping probe
Implemented a ping probe using the os ping utility. By default it sends
out 10 probes simultaneously (using preload, NOTE: more than 3 preload
requires root privileges) and returns the last line of the output which
contains the min/max/avg results.

I've left the original IcmpProbe in place for now as it provides a
template of how to send a Task across to the tcp_server. I will use this
code to for instance implement the heartbeat feature