Added node count and handling conn. lost errors
The tcp_server is now printing the amount of current connected nodes.

Added the connection_lost() function to the tcp_server protocol. If the
connection is lost due to an exception (that means not received EOF or
the connection wasn't closed by the server side) it will make sure to
unregister the node from the message_handler