Added Task result handler
the task manager now has a task result handler. It runs in a seperate
thread and checks if there is anything placed onto the task_result_queue. At the moment we're just printing the results but in the future we would want for instance to write the probe results to a database