Expanded PingProbe with error checking
If the ping command returns something on stderr this is added to the
probe result. Also added the amount of packets sent and received to the
results as well as result details when the host did not reply to any
probes.

Fixed small bug where the run_at time of persistent recurring tasks was
not updated with the recurrence_time