Started with node-frontend implementation
A very rough version of the node-frontend module was implemented. The
intention is to be able to administer the application using a web based
interface. Flask is used for the python backend and expecting to use the
bootstrap javascript library to quickly build a user interface.

At the moment the only thing we do is retrieve all present probes from
the mongodb database and print them out in an html table.

Also fixed a small issue in the output of the TraceProbe where the
min/max/avg values were not extraced properly