Reworked task_storage setup
Each task is now registered in MongoDB. If a result is returned the task
in the database is updated with the result. At the moment results are
overwritten, we should append results in the future.

It's also becoming a bit complicated creating a new task, have to see if
we can simplify it§