Adjusted node-frontend to display db tasks
The node-frontend is now displaying tasks and potential results. We're
now storing the task itself in the db as soon as it's registered by the
application. The mongoDB _id is saved in the task as task_id and if we
receive a result from that task_id we update the relevant task in
mongoDB.

There are quite a few issues to work out. We're having to pass even more
variables to the Task __init__ class. Also the overridden run()
functions are getting more complicated. The task run_at variable isn't
stored properly in mongoDB (shows 'now' instead of timestamp) and we're
not adding any additional parameters from the subclasses like dest_addr
in the Ping and TraceProbe classes.

Have to rethink how to clean this code up but for now we've got
something up and running