Writing tasks to db using new db_record function
To remove the need for having to pass the task_storage object with every
task a new function is added to the Task class called db_record. This
function returns a dictionary object containing all fields that you want
to write back to the database. The function is called in the
task_storage.add(Task) function, which in turn gets called by the task
manager as soon as a new task is put onto the task queue.

Each subclass of Task can implement their own representation this way