Reused old flask site for the netMon frontend
Copied an old flask site to be used as the netMon frontend. The site
comes with a basic authentication scheme, bootstrap.js and some
templates. A lot of references to the old site still present and the
code and file layout is not pretty (understatement). Need to do a
thorough cleanup on the code but this gives us something to start with.