Added frontend form for adding new task
The add task page will allow you to submit a new task to the database.
The backend will have to periodically retrieve new tasks from the
database and schedule them in. The backend part is not arranged yet and
submitting the form will just return you to the same add task page with
the filled in values