Added basic frontend graph to PingProbe task
Using the library python-nvd3 we're generating a ds3.js based graph in
the frontend. At the moment only the PingProbe tasks have a graph
accociated which shows the min/avg/max latency over time

Had to convert back the timestamps on the PingProbe and TraceProbe
db_result() output to unix timestamp format as this is what nvd3 takes.
We're better off having a function to convert unix timestamps to nice
time/date format in the frontend