Node (un)register now using config file details
The node task_manager now loads a config file that contains details like
name, group and version. This is used by the register and unregister
messages to let the server know who's talking.

The server is now registering and unregistering the node based on these
details. It now stores these details in self.nodes. Previously this list
only contain3ed the peername tuples. Now we're storing a dict with the
node details and the peername tuple.

Next step is to check if the name is already taken and generate a unique
name on the tcp_server if the node provided a None name value during
registration. The node should then update it's configuration file with
the generated name. This will allow you to easily add new nodes to the
environment but does require some kind of storage mechanism on the
server side. Have to decide what this storage system should be (e.g.
mongodb, mariadb, flat files, etc)