Created basic storage module for nodes
The module is used by the tcp_server to store nodes that connected to
it. When a node registers it creates an entry and sets the field
'is_connected' to True. When an unregister message is received it will
update the db record to set 'it_connected' to False.

At the moment a new record is created everytime a node connects. Have to
build in a check to see if a node with that node name already exist in
the database. If so, update the is_connected record, if not, create a
new entry.