Added Node overview frontpage and trace hop fix
The frontend now includes an overview of all known nodes in the
database. It currently displays the node name, version, group and
connection status.

Also fixed the ordering of the TraceProbe hops by using a list instead
of a dict (lists maintain order). The first hop will be the first item
in the list ([0]) and so on.