Added node removal to frontend
Nodes can now be removed from the node storage database through the
frontend home page. A delete button is present which searches for a
record in the db by nodename and removes it. The node_storage.remove
function will need error checking still (as do a lot of the other
functions btw but who's keeping score?!)