Simplified ping and trace probes
Removed support for windows from TraceProbe as well as the system
platform checking on the PingProbe. Also removing the first returned
line from the traceroute command instead of checking each line for the
presence of "traceroute to...".

These adjustments avoid us to do any calls to <str>.startswith.
Premature optimisation is the root of all evil they say... well, I say
simplification of code is the root of all happiness