Replaced traceprobe regex with ipaddress module
Removed the regex parsing of a line from the TraceProbe module. Instead
we're now splitting the string and fetching the second ([1]) item. We
check if it's a valid ip address with the python3 ipaddress module. On
value error check if we had an unresolvable hop (*). If not return None,
else return the hop