Adding task to db through frontend added
A rough version of adding a task to the database through the frontend is
added. Still have to add the run_at and probe_types to the jinja2
template add_task.html. Also have to verify if we need to store int or
strings for the various probe options. Need to build in a ip address
checker as well. We also want to add a name or description to tasks so
you can easily identify it's purpose. This will also have to be
reflected in the node code