Loading task from db periodically
In the main node loop we're fetching all probes every 5 seconds and
feeding them back into the task_manager. Next step is to make sure we
only add tasks to the queue that are not added earlier (e.g. make them
unique by name/description)

Also build in a check for existence of 'result's on probes in the
frontend to avoid https errors when loading a probe result page without
any results yet.