Tasks can be assigned to nodes or groups
Added run_on_nodes and run_on_groups lists to the Task class. It's still
a very rough implementation at the moment. The node now checks the
database every 5 seconds and adds task that should be running on his
nodename to the task_manager.

Still need to build in checks to see if the task is already scheduled.We
could accociate a unique name to Tasks to determine this and keep a list
of current scheduled tasks in the task_manager.