Adding new probes through frontend working
We can now use the add task form to actually add a new probe to a node.
It will automatically use the current time instead of prompting for a
timedate field at the moment. You have to manually type the node you wnt
to schedule it on. In the future want to provide a drop down with all
known nodes.

Need to come up with a proper way to see if a probe is already
scheduled. At the moment using a list to keep track of the dest_addr but
this isn't unique enough and still duplicates a task once since it needs
to put it on the task_list first.