Created Poller service
Created a Poller as a seperate service. The poller will be responsible
for scheduling and running all the Probes. Will remove the probe
    scheduling from the node task_manager. The task_manager will only be
    responsible for handling messages to the node controller server.

Before adding a probe to the queue it checks if it's already in the
database (based on dest_addr and run_at value). If it is it won't be
added again. Need to find out if we can solve this without having to
query the database to speed up the process