Updated Poller to accept both db and manual probes
The poller service now accepts both probes added through the frontend
and through the main event loop. The run_at value is evaluated to see if
the probe is already present in the db (which is not foulproof!!).

Somewhere along the line the rescheduling got broken. It seems that the
run_at updates by the reschedule method are not seen by the poller so
tasks get run as soon as they can.

Also made cosmetic updates to the frontend, now showing probes instead
of tasks